Integrate Swap/Add liquidity/Token creation with vpools contract
